+++ src/point_generators/bresenham_point_generator.hpp
@@ -5,15 +5,15 @@
 class BresenhamPointGenerator: public PointGenerator
 {
     std::vector<Point> generate(const Point& p1, const Point& p2) override {
-        int64_t dx = abs(int64_t(p2.x - p1.x));
-        int64_t dy = abs(int64_t(p2.y - p1.y));
+        int64_t dx = abs(p2.x - p1.x);
+        int64_t dy = abs(p2.y - p1.y);
 
         if (dx > dy) return bresenhamInnerImpl(p1.x, p1.y, p2.x, p2.y, dx, dy, 0);
         return bresenhamInnerImpl(p1.y, p1.x, p2.y, p2.x, dy, dx, 1);
     }
 
 private:
-    std::vector<Point> bresenhamInnerImpl(size_t x1, size_t y1, size_t x2, size_t y2, int64_t dx, int64_t dy, int decide) {
+    std::vector<Point> bresenhamInnerImpl(int64_t x1, int64_t y1, int64_t x2, int64_t y2, int64_t dx, int64_t dy, int decide) {
         std::vector<Point> res;
 
         int64_t pk = 2 * dy - dx;
